Add files via upload
prob1: To create String and find the sum of its length.
prob2: To check whether the number is positive or negative.
prob3: To check whether the number is positive or negative and divisible by 2 or not. 
prob4: To print the number if its Divisible by 2 and positive similarly for negative number and whole number.